refactor(brush): merge import buttons into one smart Import
Single Import button accepts both .abr and .json files. Routes to
ABR decoder or JSON preset loader based on file extension.
